+++ lib/applitools/core/session_start_info.rb
@@ -2,38 +2,36 @@
 
 module Applitools
   class SessionStartInfo
-    attr_accessor :app_id_or_name, :scenario_id_or_name
+    include Applitools::Jsonable
+    json_fields :batchInfo, :agentId, :appIdOrName, :verId, :environment, :environmentName, :branchName, :defaultMatchSettings,
+      :scenarioIdOrName, :properties, :parentBranchName, :compareWithParentBranch, :baselineEnvName, :saveDiffs, :sessionType,
+      :baselineBranchName
+
+    wrap_data do |value|
+      { startInfo: value }
+    end
 
     def initialize(options = {})
-      @agent_id = options[:agent_id]
-      @app_id_or_name = options[:app_id_or_name]
-      @ver_id = options[:ver_id]
-      @scenario_id_or_name = options[:scenario_id_or_name]
-      @batch_info = options[:batch_info]
-      @env_name = options[:env_name]
-      @environment = options[:environment]
-      @default_match_settings = options[:default_match_settings]
-      @branch_name = options[:branch_name]
-      @parent_branch_name = options[:parent_branch_name]
-      @properties = options[:properties]
-      @compare_with_parent_branch = options[:compare_with_parent_branch]
+      self.agent_id = options[:agent_id]
+      self.app_id_or_name = options[:app_id_or_name]
+      self.ver_id = options[:ver_id]
+      self.scenario_id_or_name = options[:scenario_id_or_name]
+      self.batch_info = options[:batch_info]
+      self.environment_name = options[:environment_name]
+      self.baseline_env_name = options[:baseline_env_name]
+      self.environment = options[:environment]
+      self.default_match_settings = options[:default_match_settings]
+      self.branch_name = options[:branch_name]
+      self.parent_branch_name = options[:parent_branch_name]
+      self.properties = options[:properties]
+      self.compare_with_parent_branch = options[:compare_with_parent_branch]
+      self.save_diffs = options[:save_diffs]
+      self.session_type = options[:session_type]
+      self.baseline_branch_name = options[:baseline_branch_name]
     end
 
     def to_hash
-      {
-        agent_id: @agent_id,
-        app_id_or_name: @app_id_or_name,
-        ver_id: @ver_id,
-        scenario_id_or_name: @scenario_id_or_name,
-        batch_info: @batch_info && @batch_info.to_hash,
-        env_name: @env_name,
-        environment: @environment.to_hash,
-        default_match_settings: @default_match_settings,
-        branch_name: @branch_name,
-        parent_branch_name: @parent_branch_name,
-        compare_with_parent_branch: @compare_with_parent_branch,
-        properties: @properties
-      }
+      json_data
     end
   end
 end